Improve k8s client error handling (#81)
* Improve k8s client error handling

* Update images

+++ pkg/bee/nodegroup.go
@@ -2,6 +2,7 @@ package bee
 
 import (
 	"context"
+	"errors"
 	"fmt"
 	"sort"
 	"sync"
@@ -12,6 +13,8 @@ import (
 )
 
 const nodeReadyTimeout = 3 * time.Second
+
+var errKubernetesClientNotSet = errors.New("kubernetes client not set")
 
 // NodeGroup represents group of Bee nodes
 type NodeGroup struct {
@@ -150,7 +153,7 @@ type AddressesStreamMsg struct {
 // AddressesStream returns stream of addresses of all nodes in the node group
 func (g *NodeGroup) AddressesStream(ctx context.Context) (<-chan AddressesStreamMsg, error) {
 	stopped, err := g.StoppedNodes(ctx)
-	if err != nil {
+	if err != nil && err != errKubernetesClientNotSet {
 		return nil, fmt.Errorf("stopped nodes: %w", err)
 	}
 
@@ -228,7 +231,7 @@ type BalancesStreamMsg struct {
 // BalancesStream returns stream of balances of all nodes in the cluster
 func (g *NodeGroup) BalancesStream(ctx context.Context) (<-chan BalancesStreamMsg, error) {
 	stopped, err := g.StoppedNodes(ctx)
-	if err != nil {
+	if err != nil && err != errKubernetesClientNotSet {
 		return nil, fmt.Errorf("stopped nodes: %w", err)
 	}
 
@@ -262,6 +265,10 @@ func (g *NodeGroup) BalancesStream(ctx context.Context) (<-chan BalancesStreamMs
 
 // DeleteNode deletes node from the k8s cluster and removes it from the node group
 func (g *NodeGroup) DeleteNode(ctx context.Context, name string) (err error) {
+	if g.k8s == nil {
+		return errKubernetesClientNotSet
+	}
+
 	if err := g.k8s.Delete(ctx, k8sBee.DeleteOptions{
 		Name:      name,
 		Namespace: g.cluster.namespace,
@@ -308,7 +315,7 @@ type HasChunkStreamMsg struct {
 // HasChunkStream returns stream of HasChunk requests for all nodes in the node group
 func (g *NodeGroup) HasChunkStream(ctx context.Context, a swarm.Address) (<-chan HasChunkStreamMsg, error) {
 	stopped, err := g.StoppedNodes(ctx)
-	if err != nil {
+	if err != nil && err != errKubernetesClientNotSet {
 		return nil, fmt.Errorf("stopped nodes: %w", err)
 	}
 
@@ -418,7 +425,7 @@ type OverlaysStreamMsg struct {
 // TODO: add semaphore
 func (g *NodeGroup) OverlaysStream(ctx context.Context) (<-chan OverlaysStreamMsg, error) {
 	stopped, err := g.StoppedNodes(ctx)
-	if err != nil {
+	if err != nil && err != errKubernetesClientNotSet {
 		return nil, fmt.Errorf("stopped nodes: %w", err)
 	}
 
@@ -486,7 +493,7 @@ type PeersStreamMsg struct {
 // PeersStream returns stream of peers of all nodes in the node group
 func (g *NodeGroup) PeersStream(ctx context.Context) (<-chan PeersStreamMsg, error) {
 	stopped, err := g.StoppedNodes(ctx)
-	if err != nil {
+	if err != nil && err != errKubernetesClientNotSet {
 		return nil, fmt.Errorf("stopped nodes: %w", err)
 	}
 
@@ -520,6 +527,10 @@ func (g *NodeGroup) PeersStream(ctx context.Context) (<-chan PeersStreamMsg, err
 
 // NodeReady returns node's readiness
 func (g *NodeGroup) NodeReady(ctx context.Context, name string) (ok bool, err error) {
+	if g.k8s == nil {
+		return false, errKubernetesClientNotSet
+	}
+
 	return g.k8s.Ready(ctx, k8sBee.ReadyOptions{
 		Namespace: g.cluster.namespace,
 		Name:      name,
@@ -581,7 +592,7 @@ type SettlementsStreamMsg struct {
 // SettlementsStream returns stream of settlements of all nodes in the cluster
 func (g *NodeGroup) SettlementsStream(ctx context.Context) (<-chan SettlementsStreamMsg, error) {
 	stopped, err := g.StoppedNodes(ctx)
-	if err != nil {
+	if err != nil && err != errKubernetesClientNotSet {
 		return nil, fmt.Errorf("stopped nodes: %w", err)
 	}
 
@@ -620,6 +631,10 @@ func (g *NodeGroup) Size() int {
 
 // StartNode starts new node in the node group
 func (g *NodeGroup) StartNode(ctx context.Context, name string) (err error) {
+	if g.k8s == nil {
+		return errKubernetesClientNotSet
+	}
+
 	labels := mergeMaps(g.opts.Labels, map[string]string{
 		"app.kubernetes.io/instance": name,
 	})
@@ -681,6 +696,10 @@ func (g *NodeGroup) StartNode(ctx context.Context, name string) (err error) {
 
 // StartedNodes returns list of started nodes
 func (g *NodeGroup) StartedNodes(ctx context.Context) (started []string, err error) {
+	if g.k8s == nil {
+		return []string{}, errKubernetesClientNotSet
+	}
+
 	allStarted, err := g.k8s.StartedNodes(ctx, g.cluster.namespace)
 	if err != nil {
 		return nil, fmt.Errorf("k8s started nodes: %w", err)
@@ -697,6 +716,10 @@ func (g *NodeGroup) StartedNodes(ctx context.Context) (started []string, err err
 
 // StopNode stops node by scaling down its statefulset to 0
 func (g *NodeGroup) StopNode(ctx context.Context, name string) (err error) {
+	if g.k8s == nil {
+		return errKubernetesClientNotSet
+	}
+
 	return g.k8s.Stop(ctx, k8sBee.StopOptions{
 		Name:      name,
 		Namespace: g.cluster.namespace,
@@ -705,6 +728,10 @@ func (g *NodeGroup) StopNode(ctx context.Context, name string) (err error) {
 
 // StoppedNodes returns list of stopped nodes
 func (g *NodeGroup) StoppedNodes(ctx context.Context) (stopped []string, err error) {
+	if g.k8s == nil {
+		return []string{}, errKubernetesClientNotSet
+	}
+
 	allStopped, err := g.k8s.StoppedNodes(ctx, g.cluster.namespace)
 	if err != nil {
 		return nil, fmt.Errorf("k8s stopped nodes: %w", err)
@@ -755,7 +782,7 @@ type TopologyStreamMsg struct {
 // TopologyStream returns stream of Kademlia topologies of all nodes in the node group
 func (g *NodeGroup) TopologyStream(ctx context.Context) (<-chan TopologyStreamMsg, error) {
 	stopped, err := g.StoppedNodes(ctx)
-	if err != nil {
+	if err != nil && err != errKubernetesClientNotSet {
 		return nil, fmt.Errorf("stopped nodes: %w", err)
 	}
